functions/functions/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.9.0
colorama==0.4.6
cryptography==44.0.0
deprecation==2.1.0
docstring_parser==0.16
firebase-admin==6.6.0
firebase-functions==0.1.2
Flask==3.1.0
Flask-Cors==5.0.0
functions-framework==3.8.2
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-events==0.14.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.2.0
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
shapely==2.0.6
six==1.17.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
vertexai==1.71.1
watchdog==6.0.0
Werkzeug==3.1.3
